split os_alloc and os_arena
