+++ include/nitro/os.h
@@ -8,6 +8,7 @@
 #include "types.h"
 #include "consts.h"
 #include "os_protectionRegion.h"
+#include "os_arena.h"
 #include "os_alloc.h"
 #include "os_system.h"
